requirements.txt
Flask==3.0.0
reportlab==4.0.7
Werkzeug==3.0.1
pypdf==3.17.4
Pillow>=10.0.0
pdf2image>=1.16.0
gunicorn==21.2.0
